move model_id to be managed in model_spec.py for canonical state management
